[fix] norme engine done, to do: norme client

+++ srcs/Server/Client.hpp
@@ -64,7 +64,7 @@ class Client
 		Client &		operator=( Client const & rhs );
 
 		// SETTERS
-		void	setRecv_len(int const &recv_len);
+		void	setRecv_len(size_t const &recv_len);
 		void	setFill_request(char const & fill_request);
 		void	setRequest_header_size(size_t const & rq_header_size);
 		void	setFd(int const & fd);
@@ -73,22 +73,22 @@ class Client
 		void	setHeader_readed(bool const & header_readed);
 
 		// GETTERS
-		int	& getRecv_len();
-		size_t & getRequest_header_size();
-		std::string & getFill_request();
-		bool	& getHeader_parsed();
-		bool	& getHeader_readed();
-		bool	& getIs_sendable();
-		int		& getFd();
-		Parse_request & getParse_head();
-		epoll_event & getEvents();
+		size_t &		getRecv_len();
+		size_t &		getRequest_header_size();
+		std::string &	getFill_request();
+		bool &			getHeader_parsed();
+		bool &			getHeader_readed();
+		bool &			getIs_sendable();
+		int &			getFd();
+		Parse_request &	getParse_head();
+		epoll_event &	getEvents();
 	private:
 
 		// VARIABLES
 		Client();
 		Parse_request	*_parse_head;
 		epoll_event		_events;
-		int				_recv_len;
+		size_t			_recv_len;
 		size_t			_request_header_size;
 		std::string		_fill_request;
 		bool			_header_parsed;
